Refactored code + bugfix
Refactored line comparitor class code by removing the unmatched arrays as they were redundant with matched arrays. it was more effective and simple to check for matches in matched arrays then to iterate to through unmatched. Doing this allowed me to find a bug that caused dupes. Fixed that as well